+++ CHANGELOG.md
@@ -5,6 +5,8 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
+## **WORK IN PROGRESS**
+* (theimo1221) Add handle sensor handling
 ## 0.4.13 (2024-10-12)
 * (theimo1221) Backwards compatibility